[aml] speed up blending
(texture * color) costs us an extra operation on Mali when sent as varying.
Instead, since textures use the same color for the entire render, send it as
a uniform instead.

This does not apply to fonts, since the font and its shadow are rendered
together, and may not share a color (tested and looks bad).

This increases fps anywhere heavy blending is done.