[aml] decode the smallest jpeg possible
Rather than decoding large images just to be stuffed into a small square, use
the control size as a hint for our texture size. As a bonus, smaller jpeg dct
coefficients can be used.

So for several 1080p fanart images to be shown in a gallery, we shrink the
textures and skip the big decodes. Should provide a big speedup in that case.

If we come across the same image and need a bigger source texture, we do a new
decode, always keeping the biggest around.