[aml] send original image size to large texture decoder
This finishes the work from the last commit. We need to know the full image
size so that we can guarantee that we always decode the correct image. If
we've been using a smaller copy of an image to fit a control, and we need to
bump up to a bigger size, we can stop if we hit the original.